requirements.txt
Flask==2.0.0
flasgger==0.9.5
flask-jwt-extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.27.0
SQLAlchemy==1.4.17
passlib==1.7.4
psycopg2==2.9.5 # windows
# psycopg2-binary # linux
